requirements.txt
transformers>=4.41.0
accelerate>=0.33.0
torch
sentencepiece
protobuf
pandas
tqdm
pyyaml
huggingface_hub
